requirements.txt
transformers>=4.36.0
torch>=2.0.0
pyyaml
playwright
requests
huggingface_hub>=0.19.0
llama-cpp-python>=0.2.23